Refactored tilingservice
