+++ viewer-config-persistence-helpers/src/main/java/nl/b3p/viewer/helpers/services/ServiceHelper.java
@@ -2,7 +2,6 @@ package nl.b3p.viewer.helpers.services;
 
 import nl.b3p.viewer.config.services.GeoService;
 import nl.b3p.viewer.config.services.UpdateResult;
-import nl.b3p.viewer.config.services.WMSService;
 import nl.b3p.web.WaitPageStatus;
 
 import javax.persistence.EntityManager;
@@ -11,5 +10,5 @@ import java.util.Map;
 public interface ServiceHelper {
 
     public GeoService loadServiceFromURL(String url, Map params, WaitPageStatus status, EntityManager em) throws Exception;
-    public UpdateResult updateService(EntityManager em, GeoService service);
+    public UpdateResult updateService(EntityManager em, GeoService service) throws Exception;
 }
